This is a MP4 file
The Demo of App mp4 file consists of the Demo of the Mobile_Sensor_data_transfer app. In the mobile, you will see the same versions of the app with different names as, if the same name is used the Android OS is replacing the apps with the new ones even though the version numbers are different.